Check for invalid directory

+++ announce.py
@@ -88,8 +88,9 @@ def undoAnnounceCollectionEntry(recentPostsCache: {}, \
         nickname=getNicknameFromActor(actor)
         cachedPostFilename= \
             getCachedPostFilename(baseDir,nickname,domain,postJsonObject)
-        if os.path.isfile(cachedPostFilename):
-            os.remove(cachedPostFilename)
+        if cachedPostFilename:
+            if os.path.isfile(cachedPostFilename):
+                os.remove(cachedPostFilename)
         removePostFromCache(postJsonObject,recentPostsCache)
 
         if not postJsonObject.get('type'):
@@ -144,8 +145,9 @@ def updateAnnounceCollection(recentPostsCache: {}, \
         nickname=getNicknameFromActor(actor)
         cachedPostFilename= \
             getCachedPostFilename(baseDir,nickname,domain,postJsonObject)
-        if os.path.isfile(cachedPostFilename):
-            os.remove(cachedPostFilename)
+        if cachedPostFilename:
+            if os.path.isfile(cachedPostFilename):
+                os.remove(cachedPostFilename)
         removePostFromCache(postJsonObject,recentPostsCache)
 
         if not postJsonObject.get('object'):
