+++ bookmarks.py
@@ -37,8 +37,9 @@ def undoBookmarksCollectionEntry(recentPostsCache: {}, \
     nickname=getNicknameFromActor(actor)
     cachedPostFilename= \
         getCachedPostFilename(baseDir,nickname,domain,postJsonObject)
-    if os.path.isfile(cachedPostFilename):
-        os.remove(cachedPostFilename)
+    if cachedPostFilename:
+        if os.path.isfile(cachedPostFilename):
+            os.remove(cachedPostFilename)
     removePostFromCache(postJsonObject,recentPostsCache)
 
     if not postJsonObject.get('type'):
@@ -143,8 +144,9 @@ def updateBookmarksCollection(recentPostsCache: {}, \
         nickname=getNicknameFromActor(actor)
         cachedPostFilename= \
             getCachedPostFilename(baseDir,nickname,domain,postJsonObject)
-        if os.path.isfile(cachedPostFilename):
-            os.remove(cachedPostFilename)
+        if cachedPostFilename:
+            if os.path.isfile(cachedPostFilename):
+                os.remove(cachedPostFilename)
         removePostFromCache(postJsonObject,recentPostsCache)
 
         if not postJsonObject.get('object'):
